Pull in latest from autorope

+++ scripts/profile.py
@@ -27,8 +27,8 @@ def profile(model_path, model_type):
     if "rnn" in model_type or "3d" in model_type:
         seq_len = cfg.SEQUENCE_LENGTH
 
-    #generate random array in the right shape
-    img = np.random.rand(int(h), int(w), int(ch)).astype(np.uint8)
+    # generate random array in the right shape in [0,1)
+    img = np.random.rand(int(h), int(w), int(ch))
 
     if seq_len:
         img_arr = []
@@ -36,7 +36,7 @@ def profile(model_path, model_type):
             img_arr.append(img)
         img_arr = np.array(img_arr)
 
-    #make a timer obj
+    # make a timer obj
     timer = FPSTimer()
 
     try:
